chocoq lib fix

+++ janusq/application/chocoq/chocoq/problems/k_partition_problem.py
@@ -17,7 +17,7 @@ class KPartitionProblem(LcboModel):
     def get_feasible_solution(self):
         """ 根据约束寻找到一个可行解 """
         import numpy as np
-        fsb_lst = np.zeros(len(self.variables))
+        fsb_lst = np.zeros(len(self.variables), dtype=int)
 
         t = 0
         for b, p in enumerate(self.block_allot):
